Move logging package to pkg/lib/logging
