+++ dev/operator_local.sh
@@ -76,7 +76,7 @@ fi
 export CORTEX_OPERATOR_IN_CLUSTER=false
 export CORTEX_CLUSTER_CONFIG_PATH=~/.cortex/cluster-dev.yaml
 export CORTEX_DISABLE_JSON_LOGGING=true
-export CORTEX_OPERATOR_LOG_LEVEL=debug
+export CORTEX_LOG_LEVEL=debug
 export CORTEX_PROMETHEUS_URL="http://localhost:9090"
 
 portForwardCMD="kubectl port-forward -n default prometheus-prometheus-0 9090"
